requirements.txt
Flask>=3.0.0 # Pinning major version
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31 # Trying latest patch for Python 3.13 compatibility
google-generativeai
WeasyPrint # Switched back from fpdf2
python-dotenv # To manage API keys securely
PyMySQL # Explicitly add the driver used in config
